app/requirements.txt
sklearn
pyrouge
tensorflow-gpu==1.13.1
nltk==3.4
picklable-itertools
